trackerBot/requirements.txt
requests
web3
python-telegram-bot==13.7
load_dotenv
